tweak db schema

+++ core/schema.sql
@@ -25,7 +25,7 @@ CREATE TABLE experts(
     description     TEXT              DEFAULT NULL,
     status          SMALLINT          NOT NULL,
     photo           BYTEA             DEFAULT NULL,
-    skills          SMALLINT[]        DEFAULT '{}'::SMALLINT[],
+    skills          INTEGER[]        DEFAULT '{}'::INTEGER[],
     PRIMARY KEY (user_id),
     CONSTRAINT FK_UserId FOREIGN KEY (user_id) REFERENCES users(id)
 );
